motor: updated motor interface, cleaned up ddsm210 motor driver

+++ src/control/robot_base/include/robot_base/speed_actuator_group.hpp
@@ -26,10 +26,10 @@ class SpeedActuatorGroup final : public MotorControllerInterface {
   }
 
   // public interface
-  void SetSpeed(int32_t rpm) override;
-  int32_t GetSpeed() override;
+  void SetSpeed(float rpm) override;
+  float GetSpeed() override;
 
-  void ApplyBrake(double brake) override;
+  void ApplyBrake(float brake) override;
   void ReleaseBrake() override;
 
   bool IsNormal() override;
