+++ src/interface/include/interface/driver/motor_controller_interface.hpp
@@ -18,20 +18,20 @@ class MotorControllerInterface {
   virtual ~MotorControllerInterface() = default;
 
   // public interface
-  virtual void SetSpeed(int32_t rpm) = 0;
-  virtual void SetPosition(double position) {};
-  virtual void SetTorque(double torque) {};
-  virtual void SetCurrent(double current) {};
+  virtual void SetSpeed(float rpm) = 0;
+  virtual void SetPosition(float position) {};
+  virtual void SetTorque(float torque) {};
+  virtual void SetCurrent(float current) {};
 
-  virtual int32_t GetSpeed() = 0;
+  virtual float GetSpeed() = 0;
 
-  virtual double GetPosition() { return 0; };
+  virtual float GetPosition() { return 0; };
 
-  virtual double GetTorque() { return 0; };
+  virtual float GetTorque() { return 0; };
 
-  virtual double GetCurrent() { return 0; };
+  virtual float GetCurrent() { return 0; };
 
-  virtual void ApplyBrake(double brake) {};
+  virtual void ApplyBrake(float brake) {};
   virtual void ReleaseBrake() {};
 
   virtual bool IsNormal() { return true; };
